Benchmark suite migration: decimal -> double + generator aggregate-type fix (#297)
* [WIP] Quarry.Generator: fix aggregate CLR-type resolution (partial) + session

Phase 1 of benchmark-double-migration, suspended before the typed-marker
rename per user pushback on the "object" sentinel.

What is in this commit:
- ProjectionAnalyzer.ResolveAggregateClrType reordered to consult the
  schema-driven column lookup first, then SemanticModel argument type,
  then a gated SemanticModel invocation-return-type fallback.
- 6 Sum/Avg call sites changed from the bogus "decimal" default to the
  interim "object" sentinel so ChainAnalyzer.BuildProjection's enrichment
  pass converts the unresolved type into the real column type.
- 5 new tests in AggregateTypeResolutionTests.cs covering Sum over
  double/decimal/int/long columns and Avg over double — all passing.
- Full suite remains green at 3482/3482 (baseline 3477 + 5 new).

What remains in Phase 1:
- Replace the bare "object" sentinel at the 6 aggregate call sites with
  a named TypeClassification.UnresolvedTypeMarker constant ("?"), already
  recognized by both IsUnresolvedTypeName helpers. Rename refactor only;
  behavior unchanged.

See _sessions/benchmark-double-migration/ for the full workflow state.

* [WIP] session: record WIP commit hash 892312d and session log entry

* Quarry.Generator: complete Phase 1 — typed UnresolvedTypeMarker sentinel

Introduces TypeClassification.UnresolvedTypeMarker = "?" as the canonical
sentinel for an unresolved aggregate CLR type produced by Stage 1
syntax-only analysis. Replaces the bare "object" magic-string at the 8
Sum/Avg call sites in ProjectionAnalyzer (regular, joined, window, and
joined-window aggregate paths) with the named constant, removing the
ambiguity between the legitimate "object" CLR type and an
unresolved-pending-enrichment marker.

Behavior preserving — both IsUnresolvedTypeName helpers already recognize
"?". Min/Max defaults remain at "object" (broader follow-up #1 in plan).

Pairs with the earlier reorder/gate of ResolveAggregateClrType (WIP
892312d) to close Phase 1 of benchmark-double-migration. All 5
AggregateTypeResolutionTests pass; full suite green (146 + 201 + 3143
= 3490 passed, 0 failed).

* Benchmarks: migrate Total/UnitPrice/LineTotal from decimal to double

Phase 2 of benchmark-double-migration. Switches the money-shaped columns
on OrderSchema/OrderItemSchema and their EF/DTO mirrors from decimal to
double, and updates DatabaseSetup seed literals (10.0m/1.5m/5.0m/2.5m
→ 10.0/1.5/5.0/2.5) accordingly. The DapperOrderLagDto workaround class
is removed; the regular OrderLagDto now uses double/double? so Dapper
no longer needs a parallel type.

This removes the SQLite GetDecimal(string-parse) per-cell tax from every
reader on the benchmark hot path, so the inter-library comparison
measures library overhead rather than a driver implementation choice.

Quarry.Benchmarks does NOT compile in isolation after this commit (the
reader bodies still call GetDecimal/ExecuteScalarAsync<decimal> and
three benchmark files remain parked as `.cs.disabled`). Phase 3 restores
compilation. Quarry.Tests builds and the full test suite still passes.

* Benchmarks: update readers and Aggregate/Avg signatures to double

Phase 3 of benchmark-double-migration. Restores compilation of
Quarry.Benchmarks after the Phase 2 schema/DTO migration:

- Replace `reader.GetDecimal(1)` with `reader.GetDouble(1)` across 8
  reader benchmarks (CteSimple, CteMulti, CteProjection,
  ComplexJoinFilterPaginate, JoinInner, JoinThreeTable, WindowLag,
  WindowRunningSum) — 13 call sites total.
- WindowLagBenchmarks.Dapper_Lag now uses the regular OrderLagDto;
  the DapperOrderLagDto workaround class was removed in Phase 2.
- AggregateSumBenchmarks / AggregateAvgBenchmarks switched their
  Raw/Dapper/EfCore/Quarry/SqlKata methods from `Task<decimal>` /
  `ExecuteScalarAsync<decimal>` / `Convert.ToDecimal` to `<double>` /
  `Convert.ToDouble`.

Quarry.Benchmarks builds clean.

* Benchmarks: remove obsolete GetDecimal documentation

Phase 4 of benchmark-double-migration. With the schema now using `double`
for Total/UnitPrice/LineTotal, the GetDecimal string-parse cost is no
longer on the benchmark hot path and the documentation explaining it is
stale. Drops:

- The canonical 22-line NOTE block at the top of CteSimpleBenchmarks.cs
  that documented the SqliteValueReader.GetDecimal implementation, Dapper's
  IL-emitted indexer path, and Quarry's deliberate refusal to use the
  (decimal)GetDouble trick.
- The 7 cross-reference comment blocks in ComplexJoinFilterPaginate,
  CteMulti, CteProjection, JoinInner, JoinThreeTable, WindowLag, and
  WindowRunningSum that pointed at the canonical block.

The rationale lives in PR and commit history for anyone who needs it.

* Workflow: mark Phase 5 complete, transition to REVIEW

Full test suite green in Release (3490 / 3490 passed). Benchmark
smoke run executed 30 representative benchmarks across CteSimple,
AggregateSum, and WindowLag end-to-end with no failures; Quarry
tracks the hand-rolled Raw baseline within ~1% on WindowLag
(~132µs each), matching the empirical numbers cited in the
problem statement.

All 5 implementation phases of benchmark-double-migration are
complete; advancing to REVIEW.

* REMEDIATE: extend Stage 4 enrichment to joined scalar aggregates

Addresses review findings #14 (A, low) and #10 (B, medium):

- TypeClassification.UnresolvedTypeMarker: public const → internal const
  (matches the class's accessibility).
- Add 3 unit tests for joined-aggregate, single-entity window-aggregate,
  and joined-window-aggregate paths to AggregateTypeResolutionTests.

Writing the joined tests surfaced a latent bug deeper than the original
fix scope: ProjectionAnalyzer.AnalyzeJoinedInvocation (the entry point for
`(u, o) => Sql.Sum(o.Total)` scalar joined projections) was constructing
the aggregate ProjectedColumn without setting TableAlias. Stage 4
ChainAnalyzer.BuildProjection's aggregate enrichment then called
TryResolveAggregateTypeFromSql with a null tableAlias, the alias-keyed
lookup in perAliasLookup fell through, and the unresolved marker leaked
all the way to the carrier (IJoinedQueryBuilder<User, Order, ?>) and
reader Func type — uncompilable.

The sibling ResolveJoinedAggregate (used for tuple-element joined
projections) already extracts and sets TableAlias correctly. The fix
mirrors that logic in AnalyzeJoinedInvocation — pull the alias from the
first column argument (o.Total → "t1") and pass it through to the
ProjectedColumn constructor.

Behavior preserving for all previously-working paths. Joined scalar
aggregates over non-decimal columns now compile and produce correctly-
typed readers. Full suite: 201 + 146 + 3146 = 3493 / 3493 passed.

See _sessions/benchmark-double-migration/review.md and workflow.md
decision dated 2026-05-19 for full context.

* chore: remove session artifacts before merge

+++ src/Quarry.Benchmarks/Benchmarks/AggregateAvgBenchmarks.cs
@@ -10,36 +10,36 @@ namespace Quarry.Benchmarks.Benchmarks;
 public class AggregateAvgBenchmarks : BenchmarkBase
 {
     [Benchmark(Baseline = true)]
-    public async Task<decimal> Raw_Avg()
+    public async Task<double> Raw_Avg()
     {
         await using var cmd = Connection.CreateCommand();
         cmd.CommandText = "SELECT AVG(Total) FROM orders";
         var result = await cmd.ExecuteScalarAsync();
-        return Convert.ToDecimal(result);
+        return Convert.ToDouble(result);
     }
 
     [Benchmark]
-    public async Task<decimal> Dapper_Avg()
+    public async Task<double> Dapper_Avg()
     {
-        return await Connection.ExecuteScalarAsync<decimal>("SELECT AVG(Total) FROM orders");
+        return await Connection.ExecuteScalarAsync<double>("SELECT AVG(Total) FROM orders");
     }
 
     [Benchmark]
-    public async Task<decimal> EfCore_Avg()
+    public async Task<double> EfCore_Avg()
     {
         return await EfContext.Orders.AsNoTracking().AverageAsync(o => o.Total);
     }
 
     [Benchmark]
-    public async Task<decimal> Quarry_Avg()
+    public async Task<double> Quarry_Avg()
     {
         return await QuarryDb.Orders()
             .Select(o => Sql.Avg(o.Total))
-            .ExecuteScalarAsync<decimal>();
+            .ExecuteScalarAsync<double>();
     }
 
     [Benchmark]
-    public async Task<decimal> SqlKata_Avg()
+    public async Task<double> SqlKata_Avg()
     {
         var query = new Query("orders").AsAverage("Total");
         var compiled = SqlKataCompiler.Compile(query);
@@ -51,6 +51,6 @@ public class AggregateAvgBenchmarks : BenchmarkBase
             cmd.Parameters.AddWithValue($"@p{cmd.Parameters.Count}", binding);
         }
         var result = await cmd.ExecuteScalarAsync();
-        return Convert.ToDecimal(result);
+        return Convert.ToDouble(result);
     }
 }
