+++ src/Quarry.Benchmarks/Benchmarks/JoinInnerBenchmarks.cs
@@ -9,8 +9,6 @@ using SqlKata.Compilers;
 
 namespace Quarry.Benchmarks.Benchmarks;
 
-// Reader floor here is bounded by Microsoft.Data.Sqlite.GetDecimal (string-parse path).
-// See CteSimpleBenchmarks.cs for the full explanation of why Dapper appears faster.
 public class JoinInnerBenchmarks : BenchmarkBase
 {
     [Benchmark(Baseline = true)]
@@ -25,7 +23,7 @@ public class JoinInnerBenchmarks : BenchmarkBase
             results.Add(new UserOrderDto
             {
                 UserName = reader.GetString(0),
-                Total = reader.GetDecimal(1)
+                Total = reader.GetDouble(1)
             });
         }
         return results;
@@ -85,7 +83,7 @@ public class JoinInnerBenchmarks : BenchmarkBase
             results.Add(new UserOrderDto
             {
                 UserName = reader.GetString(0),
-                Total = reader.GetDecimal(1)
+                Total = reader.GetDouble(1)
             });
         }
         return results;
