+++ src/Quarry.Benchmarks/Benchmarks/JoinThreeTableBenchmarks.cs
@@ -9,8 +9,6 @@ using SqlKata.Compilers;
 
 namespace Quarry.Benchmarks.Benchmarks;
 
-// Reader floor here is bounded by Microsoft.Data.Sqlite.GetDecimal (string-parse path).
-// See CteSimpleBenchmarks.cs for the full explanation of why Dapper appears faster.
 public class JoinThreeTableBenchmarks : BenchmarkBase
 {
     [Benchmark(Baseline = true)]
@@ -30,7 +28,7 @@ public class JoinThreeTableBenchmarks : BenchmarkBase
             results.Add(new UserOrderItemDto
             {
                 UserName = reader.GetString(0),
-                Total = reader.GetDecimal(1),
+                Total = reader.GetDouble(1),
                 ProductName = reader.GetString(2)
             });
         }
@@ -100,7 +98,7 @@ public class JoinThreeTableBenchmarks : BenchmarkBase
             results.Add(new UserOrderItemDto
             {
                 UserName = reader.GetString(0),
-                Total = reader.GetDecimal(1),
+                Total = reader.GetDouble(1),
                 ProductName = reader.GetString(2)
             });
         }
